create: to switch current personal account
