+++ web/css/clients/clients-style.css
@@ -70,7 +70,7 @@
     background-size: cover;
     background-repeat: no-repeat;
     border: none;
-    /*box-shadow: #333 0px 0px 2px 0px;*/
+    box-shadow: #333 0px 0px 2px 0px;
 }
 .container-fluid {
     padding: 20px;
@@ -101,7 +101,7 @@ li.account-info {
     width: 60px;
     height: 60px;
     border-radius: 100%;
-    border: 4px solid #fff;
+    /*border: 4px solid #fff;*/
     box-shadow: 0 0 0px 6px #9e95d7;
 }
 .navbar_user-block {
@@ -854,8 +854,8 @@ li.partners-info > img {
 .user-login-name {
     padding: 2px 20px;
     border-radius: 25px;
-    background: #598eff;
-    color: #ffffff;
+    background: #ef693a;
+    color: #fff;
     font-size: 12px;
 }
 .user-role-name {
